:wrench: clean travis install

+++ .travis/install_singularity.sh
@@ -1,13 +1,12 @@
 #!/bin/bash -ex
 
-#if [ -f "/home/travis/build/leukgen/toil_container/singularity/bin/singularity" ]; then
-    #sudo rm -rf /home/travis/build/leukgen/toil_container/singularity
+if [ -f "/home/travis/build/leukgen/toil_container/singularity/bin/singularity" ]; then
     apt-get update && apt-get install squashfs-tools dh-autoreconf build-essential
     git clone https://github.com/singularityware/singularity.git
     cd singularity
     ./autogen.sh
-    ./configure --prefix=/home/travis/build/leukgen/toil_container/singularity2
+    ./configure --prefix=/home/travis/build/leukgen/toil_container/singularity
     make install
-#else
+else
     echo exists
-#fi
+fi
